@@ -1,23 +1,21 @@
 using PddOpenSdk.AspNetCore;
-using System;
 
-namespace Microsoft.Extensions.DependencyInjection
+namespace Microsoft.Extensions.DependencyInjection;
+public static class PddServiceExtensions
 {
-    public static class PddServiceExtensions
-    {
 
-        /// <summary>
-        /// 添加批多多服务
-        /// </summary>
-        /// <param name="services"></param>
-        /// <param name="optionsAction"></param>
-        public static void AddPdd(this IServiceCollection services, Action<PddOptions> optionsAction = null)
+    /// <summary>
+    /// 添加批多多服务
+    /// </summary>
+    /// <param name="services"></param>
+    /// <param name="optionsAction"></param>
+    public static void AddPdd(this IServiceCollection services, Action<PddOptions> optionsAction = null)
+    {
+        if (optionsAction != null)
         {
-            if (optionsAction != null)
-            {
-                services.Configure(optionsAction);
-            }
-            services.AddSingleton(typeof(PddService));
+            services.Configure(optionsAction);
         }
+        services.AddSingleton(typeof(PddService));
     }
 }
+
