@@ -1,11 +1,6 @@
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
-using PddOpenSdk.Models;
-using System;
-using System.Collections.Generic;
+
 using System.Net.Http;
-using System.Text;
-using System.Threading.Tasks;
+using PddOpenSdk.Models;
 
 namespace PddOpenSdk.Services.PddApi
 {
@@ -63,7 +58,7 @@ namespace PddOpenSdk.Services.PddApi
                     dic.Add("state", state);
                 }
 
-                var data = new StringContent(JsonConvert.SerializeObject(dic), Encoding.UTF8, "application/json");
+                var data = new StringContent(JsonSerializer.Serialize(dic), Encoding.UTF8, "application/json");
                 using (var hc = new HttpClient())
                 {
                     var response = await hc.PostAsync(TokenUrl, data);
@@ -72,16 +67,16 @@ namespace PddOpenSdk.Services.PddApi
                     if (response.IsSuccessStatusCode)
                     {
                         string jsonString = await response.Content.ReadAsStringAsync();
-                        var jObject = JObject.Parse(jsonString);
-                        if (jObject.TryGetValue("error_response", out var errorResponse))
+                        var jObject = JsonDocument.Parse(jsonString);
+                        if (jObject.RootElement.TryGetProperty("error_response", out var errorResponse))
                         {
-                            ErrorResponse = JsonConvert.DeserializeObject<ErrorResponse>(jsonString);
+                            ErrorResponse = JsonSerializer.Deserialize<ErrorResponse>(jsonString);
                             Console.WriteLine("错误信息:" + errorResponse.ToString());
                             return default;
                         }
                         else
                         {
-                            var result = JsonConvert.DeserializeObject<AccessTokenResponseModel>(jsonString);
+                            var result = JsonSerializer.Deserialize<AccessTokenResponseModel>(jsonString);
                             return result;
                         }
                     }
@@ -118,13 +113,13 @@ namespace PddOpenSdk.Services.PddApi
                     dic.Add("state", state);
                 }
 
-                var data = new StringContent(JsonConvert.SerializeObject(dic), Encoding.UTF8, "application/json");
+                var data = new StringContent(JsonSerializer.Serialize(dic), Encoding.UTF8, "application/json");
                 using (var hc = new HttpClient())
                 {
                     var response = await hc.PostAsync(TokenUrl, data);
                     string jsonString = await response.Content.ReadAsStringAsync();
                     System.Console.WriteLine(jsonString);
-                    var result = JsonConvert.DeserializeObject<AccessTokenResponseModel>(jsonString);
+                    var result = JsonSerializer.Deserialize<AccessTokenResponseModel>(jsonString);
 
                     return result;
                 }
