@@ -1,20 +1,19 @@
-using Newtonsoft.Json;
 
 namespace PddOpenSdk.Models
 {
     public class AccessTokenResponseModel
     {
-        [JsonProperty("scope")]
+        [JsonPropertyName("scope")]
         public string[] Scope { get; set; }
-        [JsonProperty("access_token")]
+        [JsonPropertyName("access_token")]
         public string AccessToken { get; set; }
-        [JsonProperty("expires_in")]
+        [JsonPropertyName("expires_in")]
         public int ExpiresIn { get; set; }
-        [JsonProperty("refresh_token")]
+        [JsonPropertyName("refresh_token")]
         public string RefreshToken { get; set; }
-        [JsonProperty("owner_id")]
+        [JsonPropertyName("owner_id")]
         public string OwnerId { get; set; }
-        [JsonProperty("owner_name")]
+        [JsonPropertyName("owner_name")]
         public string OwnerName { get; set; }
     }
 }
